// src/api/priority/error/error.messages.ts
export enum PriorityMessagesError {
  PRIORITY_ERROR = 'Oops... There were problems creating or editing the priority. Please try again later',
  PRIORITY_NOT_FOUND = 'Priority not found',
  PRIORITY_ALREADY_EXIST = 'Not possible run operation, priority already exist',
  PRIORITY_NOT_EXIST = 'Priority not exists',
  INTERNAL_SERVER_ERROR = 'Internal server error. Contact with the administrator',
  INVALID_STATUS = 'Status is invalid',
  PRIORITY_NOT_FOUND_OR_NOT_ACTIVE = 'Priority not found or not active',
  PRIORITY_NOT_DELETED = 'Oops... There was an error deleting the priority. Please try again later.',
  PRIORITY_NOT_RESTORED = 'Oops... There was an error restoring the priority. Please try again later.',
  PRIORITY_NOT_UPDATED = 'Oops... There was an error updating the priority. Please try again later.',
  PRIORITY_NOT_SAVED = 'Ups... Hubo un error al crear o editar la categoría. Por favor, inténtelo de nuevo más tarde',
}