// src/api/priority/messages/priority.messages.ts
export enum PriorityMessages {
  PRIORITY_CREATED = 'Prioridad creada con éxito',
  PRIORITY_UPDATED = 'Prioridad actualizada con éxito',
  PRIORITY_ACTIVED = 'Prioridad activada con éxito',
  PRIORITY_DESACTIVED = 'Prioridad desactivada con éxito',
  PRIORITY_REMOVED = 'Prioridad eliminada con éxito',
  PRIORITY_RESTORED = 'Prioridad restaurada con éxito',
}